Refactor extensions into domain-specific classes
