+++ Patches/ServantMenuMapperPatch.cs
@@ -4,6 +4,7 @@ using ProjectM.UI;
 using Stunlock.Core;
 using Unity.Entities;
 using UnityEngine;
+using Eclipse.Utilities.Extensions;
 
 namespace Eclipse.Patches;
