+++ Utilities/Extensions/CollectionExtensions.cs
@@ -1,0 +1,54 @@
+using System;
+using System.Collections.Generic;
+
+namespace Eclipse.Utilities.Extensions;
+
+internal static class CollectionExtensions
+{
+    public static void ForEach<T>(this IEnumerable<T> source, Action<T> action)
+    {
+        foreach (var item in source)
+        {
+            action(item);
+        }
+    }
+
+    public static Dictionary<TValue, TKey> Reverse<TKey, TValue>(this IDictionary<TKey, TValue> source)
+    {
+        var reversed = new Dictionary<TValue, TKey>();
+
+        foreach (var kvp in source)
+        {
+            reversed[kvp.Value] = kvp.Key;
+        }
+
+        return reversed;
+    }
+
+    public static bool Equals<T>(this T value, params T[] options)
+    {
+        foreach (var option in options)
+        {
+            if (value.Equals(option)) return true;
+        }
+
+        return false;
+    }
+
+    public static bool TryGetMatch(this HashSet<(ulong, ulong)> hashSet, ulong value, out (ulong, ulong) matchingPair)
+    {
+        matchingPair = default;
+
+        foreach (var pair in hashSet)
+        {
+            if (pair.Item1 == value || pair.Item2 == value)
+            {
+                matchingPair = pair;
+
+                return true;
+            }
+        }
+
+        return false;
+    }
+}
