+++ Utilities/Extensions/EntityExtensions.cs
@@ -5,33 +5,31 @@ using ProjectM.Gameplay.Systems;
 using ProjectM.Network;
 using ProjectM.Scripting;
 using ProjectM.Shared;
-using Stunlock.Core;
 using System.Collections;
 using System.Runtime.InteropServices;
 using Unity.Entities;
 using Unity.Mathematics;
 using Unity.Transforms;
 using UnityEngine;
-using static Eclipse.Services.LocalizationService;
 
-namespace Eclipse;
-internal static class Extensions
+namespace Eclipse.Utilities.Extensions;
+
+internal static class EntityExtensions
 {
     static EntityManager EntityManager => Core.EntityManager;
     static ClientGameManager ClientGameManager => Core.ClientGameManager;
     static SystemService SystemService => Core.SystemService;
     static PrefabCollectionSystem PrefabCollectionSystem => SystemService.PrefabCollectionSystem;
 
-    const string EMPTY_KEY = "LocalizationKey.Empty";
-
     public delegate void WithRefHandler<T>(ref T item);
+
     public static void With<T>(this Entity entity, WithRefHandler<T> action) where T : struct
     {
         T item = entity.ReadRW<T>();
         action(ref item);
-
         EntityManager.SetComponentData(entity, item);
     }
+
     public static void AddWith<T>(this Entity entity, WithRefHandler<T> action) where T : struct
     {
         if (!entity.Has<T>())
@@ -41,6 +39,7 @@ internal static class Extensions
 
         entity.With(action);
     }
+
     public unsafe static void Write<T>(this Entity entity, T componentData) where T : struct
     {
         ComponentType componentType = new(Il2CppType.Of<T>());
@@ -54,6 +53,7 @@ internal static class Extensions
             EntityManager.SetComponentDataRaw(entity, typeIndex, byteData, size);
         }
     }
+
     static byte[] StructureToByteArray<T>(T structure) where T : struct
     {
         int size = Marshal.SizeOf(structure);
@@ -67,6 +67,7 @@ internal static class Extensions
 
         return byteArray;
     }
+
     unsafe static T ReadRW<T>(this Entity entity) where T : struct
     {
         ComponentType componentType = new(Il2CppType.Of<T>());
@@ -75,6 +76,7 @@ internal static class Extensions
         void* componentData = EntityManager.GetComponentDataRawRW(entity, typeIndex);
         return Marshal.PtrToStructure<T>(new IntPtr(componentData));
     }
+
     public unsafe static T Read<T>(this Entity entity) where T : struct
     {
         ComponentType componentType = new(Il2CppType.Of<T>());
@@ -83,6 +85,7 @@ internal static class Extensions
         void* componentData = EntityManager.GetComponentDataRawRO(entity, typeIndex);
         return Marshal.PtrToStructure<T>(new IntPtr(componentData));
     }
+
     public static unsafe bool TryGetBuffer<T>(this Entity entity, out DynamicBuffer<T> dynamicBuffer) where T : struct
     {
         if (GameManager_Shared.TryGetBuffer(EntityManager, entity, out dynamicBuffer))
@@ -93,30 +96,37 @@ internal static class Extensions
         dynamicBuffer = default;
         return false;
     }
+
     public static DynamicBuffer<T> ReadBuffer<T>(this Entity entity) where T : struct
     {
         return EntityManager.GetBuffer<T>(entity);
     }
+
     public static DynamicBuffer<T> AddBuffer<T>(this Entity entity) where T : struct
     {
         return EntityManager.AddBuffer<T>(entity);
     }
+
     public unsafe static void* GetComponentData(this Entity entity, TypeIndex typeIndex)
     {
         return EntityManager.GetComponentDataRawRO(entity, typeIndex);
     }
+
     public unsafe static void SetComponentData(this Entity entity, TypeIndex typeIndex, void* byteData, int size)
     {
         EntityManager.SetComponentDataRaw(entity, typeIndex, byteData, size);
     }
+
     public unsafe static void* GetBufferData(this Entity entity, TypeIndex typeIndex)
     {
         return EntityManager.GetBufferRawRO(entity, typeIndex);
     }
+
     public static int GetBufferLength(this Entity entity, TypeIndex typeIndex)
     {
         return EntityManager.GetBufferLength(entity, typeIndex);
     }
+
     public static void SetBufferData<T>(Entity prefabSource, T[] bufferArray) where T : struct
     {
         DynamicBuffer<T> buffer = prefabSource.Has<T>() ? prefabSource.ReadBuffer<T>() : prefabSource.AddBuffer<T>();
@@ -127,6 +137,7 @@ internal static class Extensions
             buffer.Add(element);
         }
     }
+
     public static bool TryGetComponent<T>(this Entity entity, out T componentData) where T : struct
     {
         componentData = default;
@@ -134,12 +145,12 @@ internal static class Extensions
         if (entity.Has<T>())
         {
             componentData = entity.Read<T>();
-
             return true;
         }
 
         return false;
     }
+
     public static bool TryGetComponentObject<T>(this Entity entity, EntityManager entityManager, out T componentObject) where T : class
     {
         componentObject = default;
@@ -152,48 +163,38 @@ internal static class Extensions
 
         return false;
     }
+
     public static bool TryRemoveComponent<T>(this Entity entity) where T : struct
     {
         if (entity.Has<T>())
         {
             entity.Remove<T>();
-
             return true;
         }
 
         return false;
     }
+
     public static bool Has<T>(this Entity entity)
     {
         return EntityManager.HasComponent(entity, new(Il2CppType.Of<T>()));
     }
+
     public static bool Has(this Entity entity, ComponentType componentType)
     {
         return EntityManager.HasComponent(entity, componentType);
     }
-    public static string GetPrefabName(this PrefabGUID prefabGUID)
-    {
-        return PrefabGuidsToNames.TryGetValue(prefabGUID, out string prefabName) ? $"{prefabName} {prefabGUID}" : "String.Empty";
-    }
-    public static string GetLocalizedName(this PrefabGUID prefabGUID)
-    {
-        string localizedName = GetNameFromGuidString(GetGuidString(prefabGUID));
 
-        if (!string.IsNullOrEmpty(localizedName))
-        {
-            return localizedName;
-        }
-
-        return EMPTY_KEY;
-    }
     public static void Add<T>(this Entity entity)
     {
         if (!entity.Has<T>()) EntityManager.AddComponent(entity, new(Il2CppType.Of<T>()));
     }
+
     public static void Remove<T>(this Entity entity)
     {
         if (entity.Has<T>()) EntityManager.RemoveComponent(entity, new(Il2CppType.Of<T>()));
     }
+
     public static bool TryGetFollowedPlayer(this Entity entity, out Entity player)
     {
         player = Entity.Null;
@@ -208,6 +209,7 @@ internal static class Extensions
 
         return false;
     }
+
     public static bool TryGetPlayer(this Entity entity, out Entity player)
     {
         player = Entity.Null;
@@ -215,12 +217,12 @@ internal static class Extensions
         if (entity.Has<PlayerCharacter>())
         {
             player = entity;
-
             return true;
         }
 
         return false;
     }
+
     public static bool IsPlayer(this Entity entity)
     {
         if (entity.Has<VampireTag>())
@@ -230,6 +232,7 @@ internal static class Extensions
 
         return false;
     }
+
     public static bool IsDifferentPlayer(this Entity entity, Entity target)
     {
         if (entity.IsPlayer() && target.IsPlayer() && !entity.Equals(target))
@@ -239,6 +242,7 @@ internal static class Extensions
 
         return false;
     }
+
     public static bool IsFollowingPlayer(this Entity entity)
     {
         if (entity.TryGetComponent(out Follower follower))
@@ -251,18 +255,22 @@ internal static class Extensions
 
         return false;
     }
+
     public static Entity GetBuffTarget(this Entity entity)
     {
         return CreateGameplayEventServerUtility.GetBuffTarget(EntityManager, entity);
     }
+
     public static Entity GetPrefabEntity(this Entity entity)
     {
         return ClientGameManager.GetPrefabEntity(entity.Read<PrefabGUID>());
     }
+
     public static Entity GetSpellTarget(this Entity entity)
     {
         return CreateGameplayEventServerUtility.GetSpellTarget(EntityManager, entity);
     }
+
     public static bool TryGetTeamEntity(this Entity entity, out Entity teamEntity)
     {
         teamEntity = Entity.Null;
@@ -274,13 +282,13 @@ internal static class Extensions
             if (teamReferenceEntity.Exists())
             {
                 teamEntity = teamReferenceEntity;
-
                 return true;
             }
         }
 
         return false;
     }
+
     public static bool Exists(this Entity entity)
     {
         return entity.HasValue() && entity.IndexWithinCapacity() && EntityManager.Exists(entity);
@@ -288,6 +296,7 @@ internal static class Extensions
 
     const string PREFIX = "Entity(";
     const int LENGTH = 7;
+
     public static bool IndexWithinCapacity(this Entity entity)
     {
         string entityStr = entity.ToString();
@@ -304,33 +313,30 @@ internal static class Extensions
         int closeRel = tail.IndexOf(')');
         if (closeRel <= 0) return false;
 
-        // Parse numbers
         if (!int.TryParse(span[..colon], out int index)) return false;
         if (!int.TryParse(tail[..closeRel], out _)) return false;
 
-        // Single unsigned capacity check
         int capacity = EntityManager.EntityCapacity;
         bool isValid = (uint)index < (uint)capacity;
 
-        if (!isValid)
-        {
-            // Core.Log.LogWarning($"Entity index out of range! ({index}>{capacity})");
-        }
-
         return isValid;
     }
+
     public static bool IsDisabled(this Entity entity)
     {
         return entity.Has<Disabled>();
     }
+
     public static bool HasConnectedCoffin(this Entity entity)
     {
         return entity.TryGetComponent(out ServantConnectedCoffin servantConnectedCoffin) && servantConnectedCoffin.CoffinEntity.GetEntityOnServer().Exists();
     }
+
     public static bool IsVBlood(this Entity entity)
     {
         return entity.Has<VBloodUnit>();
     }
+
     public static ulong GetSteamId(this Entity entity)
     {
         if (entity.TryGetComponent(out PlayerCharacter playerCharacter))
@@ -344,6 +350,7 @@ internal static class Extensions
 
         return 0;
     }
+
     public static NetworkId GetNetworkId(this Entity entity)
     {
         if (entity.TryGetComponent(out NetworkId networkId))
@@ -353,36 +360,31 @@ internal static class Extensions
 
         return NetworkId.Empty;
     }
-    public static void ForEach<T>(this IEnumerable<T> source, Action<T> action)
-    {
-        foreach (var item in source)
-        {
-            action(item);
-        }
-    }
+
     public static PrefabGUID GetPrefabGUID(this Entity entity)
     {
         if (entity.TryGetComponent(out PrefabGUID prefabGUID)) return prefabGUID;
-
         return PrefabGUID.Empty;
     }
+
     public static Entity GetUserEntity(this Entity character)
     {
         if (character.TryGetComponent(out PlayerCharacter playerCharacter)) return playerCharacter.UserEntity;
-
         return Entity.Null;
     }
+
     public static User GetUser(this Entity entity)
     {
         if (entity.TryGetComponent(out PlayerCharacter playerCharacter) && playerCharacter.UserEntity.TryGetComponent(out User user)) return user;
         else if (entity.TryGetComponent(out user)) return user;
-
         return User.Empty;
     }
+
     public static bool HasBuff(this Entity entity, PrefabGUID buffPrefabGUID)
     {
         return GameManager_Shared.HasBuff(EntityManager, entity, buffPrefabGUID.ToIdentifier());
     }
+
     public static bool TryGetBuff(this Entity entity, PrefabGUID buffPrefabGUID, out Entity buffEntity)
     {
         if (GameManager_Shared.TryGetBuff(EntityManager, entity, buffPrefabGUID.ToIdentifier(), out buffEntity))
@@ -392,6 +394,7 @@ internal static class Extensions
 
         return false;
     }
+
     public static float3 GetAimPosition(this Entity entity)
     {
         if (entity.TryGetComponent(out EntityInput entityInput))
@@ -401,6 +404,7 @@ internal static class Extensions
 
         return float3.zero;
     }
+
     public static bool TryGetPosition(this Entity entity, out float3 position)
     {
         position = float3.zero;
@@ -408,12 +412,12 @@ internal static class Extensions
         if (entity.TryGetComponent(out Translation translation))
         {
             position = translation.Value;
-
             return true;
         }
 
         return false;
     }
+
     public static float3 GetPosition(this Entity entity)
     {
         if (entity.TryGetComponent(out Translation translation))
@@ -423,22 +427,7 @@ internal static class Extensions
 
         return float3.zero;
     }
-    public static bool TryGetMatch(this HashSet<(ulong, ulong)> hashSet, ulong value, out (ulong, ulong) matchingPair)
-    {
-        matchingPair = default;
 
-        foreach (var pair in hashSet)
-        {
-            if (pair.Item1 == value || pair.Item2 == value)
-            {
-                matchingPair = pair;
-
-                return true;
-            }
-        }
-
-        return false;
-    }
     public static bool IsCustomSpawned(this Entity entity)
     {
         if (entity.TryGetComponent(out IsMinion isMinion) && isMinion.Value)
@@ -448,10 +437,12 @@ internal static class Extensions
 
         return false;
     }
+
     public static void Destroy(this Entity entity)
     {
         if (entity.Exists()) DestroyUtility.Destroy(EntityManager, entity);
     }
+
     public static void SetTeam(this Entity entity, Entity teamSource)
     {
         if (entity.Has<Team>() && entity.Has<TeamReference>() && teamSource.TryGetComponent(out Team sourceTeam) && teamSource.TryGetComponent(out TeamReference sourceTeamReference))
@@ -470,6 +461,7 @@ internal static class Extensions
             });
         }
     }
+
     public static void SetFaction(this Entity entity, PrefabGUID factionPrefabGUID)
     {
         if (entity.Has<FactionReference>())
@@ -480,40 +472,14 @@ internal static class Extensions
             });
         }
     }
+
     public static bool HasValue(this Entity entity)
     {
         return entity != Entity.Null;
     }
+
     public static bool IsAllied(this Entity entity, Entity player)
     {
         return ClientGameManager.IsAllies(entity, player);
-    }
-    public static Coroutine Start(this IEnumerator routine)
-    {
-        return Core.StartCoroutine(routine);
-    }
-    public static void Stop(this Coroutine routine)
-    {
-        if (routine != null) Core.StopCoroutine(routine);
-    }
-    public static Dictionary<TValue, TKey> Reverse<TKey, TValue>(this IDictionary<TKey, TValue> source)
-    {
-        var reversed = new Dictionary<TValue, TKey>();
-
-        foreach (var kvp in source)
-        {
-            reversed[kvp.Value] = kvp.Key;
-        }
-
-        return reversed;
-    }
-    public static bool Equals<T>(this T value, params T[] options)
-    {
-        foreach (var option in options)
-        {
-            if (value.Equals(option)) return true;
-        }
-
-        return false;
     }
 }
