+++ Utilities/Extensions/UiExtensions.cs
@@ -1,0 +1,40 @@
+using ProjectM;
+using System.Collections;
+using UnityEngine;
+using static Eclipse.Services.LocalizationService;
+
+namespace Eclipse.Utilities.Extensions;
+
+internal static class UiExtensions
+{
+    const string EMPTY_KEY = "LocalizationKey.Empty";
+
+    public static string GetPrefabName(this PrefabGUID prefabGUID)
+    {
+        return PrefabGuidsToNames.TryGetValue(prefabGUID, out string prefabName)
+            ? $"{prefabName} {prefabGUID}"
+            : "String.Empty";
+    }
+
+    public static string GetLocalizedName(this PrefabGUID prefabGUID)
+    {
+        string localizedName = GetNameFromGuidString(GetGuidString(prefabGUID));
+
+        if (!string.IsNullOrEmpty(localizedName))
+        {
+            return localizedName;
+        }
+
+        return EMPTY_KEY;
+    }
+
+    public static Coroutine Start(this IEnumerator routine)
+    {
+        return Core.StartCoroutine(routine);
+    }
+
+    public static void Stop(this Coroutine routine)
+    {
+        if (routine != null) Core.StopCoroutine(routine);
+    }
+}
